+++ CHANGES.md
@@ -256,6 +256,7 @@
 - Quick Scan: skip all by default (please edit `config.ini` to enable each scan)
 
 ### Pre-release (work-in-progress)
+- Spoof User-Agent string
 - HIMSA: add archived version v3.5.0.0 for Noahlink Wireless firmware upgrader
 - HIMSA: workaround for error checking file data stream
 - Quick Scan: add Starkey PatientBase and HIMSA Noahlink Wireless Firmware Upgrader