+++ CHANGES.md
@@ -255,7 +255,7 @@
 ### v2025.06.10
 - Quick Scan: skip all by default (please edit `config.ini` to enable each scan)
 
-### Pre-release (work-in-progress)
+### v2025.07.04
 - Spoof User-Agent string
 - HIMSA: add archived version v3.5.0.0 for Noahlink Wireless firmware upgrader
 - HIMSA: workaround for error checking file data stream